+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-20T21:44:12.405Z",
+    "2026-02-20T22:08:06.719Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {