+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -148,7 +148,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-20T21:44:16.567Z",
+    "2026-02-20T22:08:10.962Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {